- Created error handler for 404 error code. - Added template blank for 404 error handler. - Moved monitor request from general to gpio-service routing library. - Added forward-slash index redirect for index page. - Removed react and babel libraries, since we're not going to use them.
