+++ web/routes/gpio_service.py
@@ -3,15 +3,32 @@ Routing library for all GPIO-related requests
 """
 import json
 
+from flask import render_template
 from flask_security import login_required, roles_required, roles_accepted
 
 import config
+from config import WEB_SERVICE_CONFIG
 from board_controller.server import server_interface
 from web import web_service, utils
 from web.utils import MimeType
 
 # Until Home page and device selection is available, we hard-code it to a single configured client
 BOARD_ID = config.CLIENT_CONFIG["device_id"]
+
+
+@web_service.route("/monitor/<string:device_id>", methods=["GET"])
+@login_required
+@roles_accepted("admin")
+def monitor(device_id):
+    """
+    Maps monitor path to a static index file
+    :param device_id: Device ID
+    :return: Rendered template
+    """
+    return render_template("gpio/monitor.html",
+                           pins_config=server_interface.get_device_pin_config(device_id),
+                           socket_url=WEB_SERVICE_CONFIG["host"],
+                           socket_port=WEB_SERVICE_CONFIG["socket_port"])
 
 
 @web_service.route("/gpio/status", methods=["GET"])
